refactor(connectors): extract shared utilities and split umami-tools monolith
- Extract isRecord, getString, getBoolean and other parsing helpers
  into connector-values.ts for reuse across connector modules.
- Split umami-tools.ts (~800 lines) into umami-tool-definitions.ts
  (schemas, constants, protocol version) and umami-tool-executor.ts
  (validation, handler registry, execution).
- Replace monolithic switch with typed UMAMI_TOOL_HANDLERS registry
  to follow Open/Closed Principle.
- Parametrize embedded connector test handlers (Zendesk/Umami) via
  buildEmbeddedConnectorTestHandler factory in test-connection.ts.
